+++ bootique-mybatis/src/test/java/io/bootique/mybatis/testpojos1/TO1.java
@@ -1,0 +1,59 @@
+/*
+ * Licensed to ObjectStyle LLC under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ObjectStyle LLC licenses
+ * this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.bootique.mybatis.testpojos1;
+
+public class TO1 {
+
+    private int c1;
+    private String c2;
+    private String c3;
+    private String c3_x;
+
+    public int getC1() {
+        return c1;
+    }
+
+    public String getC2() {
+        return c2;
+    }
+
+    public String getC3() {
+        return c3;
+    }
+
+    public void setC1(int c1) {
+        this.c1 = c1;
+    }
+
+    public void setC2(String c2) {
+        this.c2 = c2;
+    }
+
+    public void setC3(String c3) {
+        this.c3 = c3;
+    }
+
+    public String getC3_x() {
+        return c3_x;
+    }
+
+    public void setC3_x(String c3_x) {
+        this.c3_x = c3_x;
+    }
+}